refactor(restaurantStorage): lib 폴더로 이동
