+++ src/components/main/Main.ts
@@ -2,7 +2,7 @@ import { Category, SortBy, Restaurant } from '../../type';
 import Component from '../Component';
 import RestaurantFilterContainer from './restaurant/RestaurantFilterContainer';
 import RestaurantList from './restaurant/RestaurantList';
-import store from '../../domain/restaurantsStorage';
+import store from '../../lib/restaurantsStorage';
 import { DEFAULT_CATEGORY } from '../../utils/constants';
 
 interface MainProps {}
